requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
firebase-admin==6.5.0
PyJWT==2.8.0
PyMySQL==1.1.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.133.0
google-cloud-storage==2.17.0
google-cloud-tasks==2.16.3
google-cloud-firestore==2.16.0
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.24.0
user-agents==2.2.0
requests==2.32.3
pytz==2024.1
python-dateutil==2.9.0
sendgrid==6.11.0
reportlab==4.2.0
XlsxWriter==3.2.0
jaconv==0.3.4
gspread==6.1.2
gspread-formatting==1.2.0
Werkzeug==3.0.3
